Update starter_code.ipynb
Made a lot of changes. Added function vote_count() to reduce the # of times we need to walk through the plan for our stats functions. This may be combined with identify boundary nodes if we have an efficiency problem in the future. Updated Stats functions to reflect the vote_counter() function. Finished format_chain_data() seems to be working. Finished the chain_run function by adding the probability component. Major overhaul on documentation.